+++ Application.MainTests/PositionTests.cs
@@ -10,17 +10,17 @@ namespace Application.MainTests
         [Test]
         public void CalculateCurrentPositionPositiveMarginTest()
         {
-            var purchasePrice = (decimal)77;
-            var closePrice = (decimal)78.52;
+            var purchasePrice = 77.00;
+            var closePrice = 78.52;
             var contractUnit = 500;
             var noOfContracts = 2.5;
-            var exchangeRate = (decimal)0.77;
+            var exchangeRate = 0.77;
 
             var trade = new TestTrade();
             trade.EntryPrice = purchasePrice;
             trade.ClosePrice = closePrice;
             trade.Direction = PositionType.BUY.ToString();
-            trade.Volume = (decimal)noOfContracts;
+            trade.Volume = noOfContracts;
 
             var margin = Margin.Calculate(contractUnit, exchangeRate, trade, closePrice, trade.Volume);
             Console.WriteLine("Margin Anticipated: £1463, Margin Actual: £" + margin);
