+++ Application/Business/BackTest/Position/IStopLossCreator.cs
@@ -7,17 +7,17 @@ namespace PikUpStix.Trading.Forecast
 {
     public interface IStopLossCreator
     {
-        decimal NewPositionStopLoss { get; set; }
-        List<TestTrade> CalculateStops(List<TestTrade> trades, decimal stopLossPercentage, decimal currentMargin, decimal volume, decimal exchangeRate, PositionValue weightedProposedPosition);
+        double NewPositionStopLoss { get; set; }
+        List<TestTrade> CalculateStops(List<TestTrade> trades, double stopLossPercentage, double currentMargin, double volume, double exchangeRate, PositionValue weightedProposedPosition);
     }
 
     public class SimpleStopLossCreator : IStopLossCreator
     {
         // Works out stop loss based on single percentage input and does not consider other position that may be on the same instrument
-        public decimal NewPositionStopLoss { get; set; }
+        public double NewPositionStopLoss { get; set; }
 
-        public List<TestTrade> CalculateStops(List<TestTrade> trades, decimal stopLossPercentage, decimal currentMargin, 
-            decimal volume, decimal exchangeRate, PositionValue weightedProposedPosition)
+        public List<TestTrade> CalculateStops(List<TestTrade> trades, double stopLossPercentage, double currentMargin, 
+            double volume, double exchangeRate, PositionValue weightedProposedPosition)
         {
             var stopLoss = new StopLoss(currentMargin,
                 stopLossPercentage,
@@ -35,10 +35,10 @@ namespace PikUpStix.Trading.Forecast
     public class AggregatedStopLossCreator : IStopLossCreator
     {
         // Works out stop loss for all positions on a given instrument to create overall risk based on StopLossPercent
-        public decimal NewPositionStopLoss { get; set; }
+        public double NewPositionStopLoss { get; set; }
 
-        public List<TestTrade> CalculateStops(List<TestTrade> trades, decimal stopLossPercentage, decimal currentMargin,
-            decimal volume, decimal exchangeRate, PositionValue weightedProposedPosition)
+        public List<TestTrade> CalculateStops(List<TestTrade> trades, double stopLossPercentage, double currentMargin,
+            double volume, double exchangeRate, PositionValue weightedProposedPosition)
         {
             var totalVolume = volume;
             if(trades.Any(x => x.InstrumentId == weightedProposedPosition.Instrument.Id))
